+++ publish/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## V 0.1.7
 - Fixed log spam when no Backrooms were generated
+- Fixed an issue where the Bracken would be stuck when holding a player due to no path to its favorit room being avalible
 
 ## V 0.1.6
 - Fixed minor code issues