+++ publish/CHANGELOG.md
@@ -1,7 +1,9 @@
 ## V 0.1.7
+- The mod is now host only
+- Fixed log spam with info that the favorite spot was changed
 - Fixed log spam when no Backrooms were generated
 - Fixed an issue where the Bracken would be stuck when holding a player due to no path to its favorite room being avalible
-- Fixed Backrroms being moved instead of the favorite position of the bracken
+- Fixed Backrooms being moved instead of the favorite position of the bracken
 
 ## V 0.1.6
 - Fixed minor code issues